Moved scaled_down_geobox to _base.py
avoid circular dependencies
